Move round_up_to_mask and friends to own header
I want to use these functions in tasks-fifo.c, so I moved them to a
runtime header where they are more available.
